+++ CHANGELOG.md
@@ -12,10 +12,13 @@
 * Rows in a `RowsBlock` will now break and continue on the next page
   when they do not fit. To prevent this, set `breakInside` to `avoid`.
 
+* Lines in a `TextBlock` will now break and continue on the next page
+  when they do not fit. To prevent this, set `breakInside` to `avoid`.
+
 ### Added
 
-* Attribute `breakInside` to control page breaks inside a rows block.
-  The default is `auto`.
+* Attribute `breakInside` to control page breaks inside a block on
+  `TextBlock` and `RowsBlock`. The default is `auto`.
 
 ## [0.4.2] - 2023-04-29